refactor(plugin-oracle): extract TLS builder into dedicated OracleSSLMapping
